fix: remove previous downloads fixtures and rename file to exported downloadCategories
