+++ src/node-lib/curriculum-api-2023/fixtures/downloadCategories.fixture.ts
@@ -1,0 +1,11 @@
+export const DOWNLOAD_CATEGORIES = [
+  "EYFS",
+  "KS1",
+  "KS2",
+  "KS3",
+  "KS4",
+  "Therapies",
+  "Specialist",
+] as const;
+
+export type DownloadCategory = (typeof DOWNLOAD_CATEGORIES)[number];
